test: merge all ApiDeployer tests
This allows running the same tests for any kind of API
